cloud_function/Product-Text-Scraper/functions/requirements.txt
functions-framework==3.*
google-cloud-storage>=2.17.0
requests>=2.32.3
pytz>=2024.1
